requirements.txt
fastapi
certifi
charset-normalizer
idna
python-dotenv
requests
urllib3
uvicorn
openai